Not cloning repo but copy it to container
Until now, the Docker image cloned the git repo to the container. This conflicts with best practices and takes the advantage of building a Docker image from forks or other branches. That is why I changed the Dockerfile to copying the source code to the container instead cloning it inside.

Another benefit is that git can be removed as dependency.