Adding the ability to connect to VPN containers
YouTube has a problem with too many requests. If you want to bulk download all your playlists, you should use VPNs. 

In order to make this possible, I added the functionality that the script queries all running VPN containers (containers that are healthy and have VPN in the name) and start the Savify containers with attached network.