Changing default location for m3u files
Defining "%playlist%/%playlist name%.m3u" as default location for m3u files is a problem when using a different grouping. The default location should be in the root download directory, making it independent of grouping and creating a standard for relative paths inside the m3u file.

Since creation of m3u playlists offers the ability to switch the grouping from "%playlist%" to "%artist%/%album%" as default (because most media servers expect this structure), having a "%playlist%" grouping is not needed anymore. Information regarding playlists are now kept in the m3u file like it should be.