Download requested format instead of lossy compression
With this change, `savify` will instruct `youtube-dl` to prefer a download in the requested format, if available. For example, if the user requests `-f opus`, `youtube-dl` will download the opus original of the audio track, instead of downloading the `m4a` one and using `ffmpeg` for a lossy conversion to `opus`.